Combined all 3 branchs into one for testing purposes
I ran into 2 errors, other than that the program compiled fine.

Jeremy - in Drint(), swapper(JOBTABLE[jobIndex].getJobNumber()); //
Gives me error: Job in already in core.

Flo - in dispatcher() , a and p needs to be passed in. It gives me error
stating that a and p are not found in the scope.